Revert "Merge remote-tracking branch 'origin/master' into Fix_Munkalap_milyen_adat_szerkesztheto_#220"
This reverts commit a74ad72f0ac44f29a73bbc29747c0e6188a97153.
