+++ backend/src/main/java/hu/flowacademy/worksheet/configuration/SpringSecurityAuditorAware.java
@@ -1,0 +1,36 @@
+package hu.flowacademy.worksheet.configuration;
+
+import hu.flowacademy.worksheet.entity.User;
+import hu.flowacademy.worksheet.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.keycloak.KeycloakPrincipal;
+import org.keycloak.KeycloakSecurityContext;
+import org.keycloak.representations.IDToken;
+import org.springframework.data.domain.AuditorAware;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Optional;
+
+@Slf4j
+@Transactional
+@RequiredArgsConstructor
+public class SpringSecurityAuditorAware implements AuditorAware<User> {
+
+    private final UserRepository userRepository;
+
+    @Override
+    public Optional<User> getCurrentAuditor() {
+        return Optional.ofNullable(SecurityContextHolder.getContext())
+                .map(SecurityContext::getAuthentication)
+                .filter(Authentication::isAuthenticated)
+                .map(authentication -> (KeycloakPrincipal) authentication.getPrincipal())
+                .map(KeycloakPrincipal::getKeycloakSecurityContext)
+                .map(KeycloakSecurityContext::getToken)
+                .map(IDToken::getEmail)
+                .flatMap(userRepository::findFirstByEmail);
+    }
+}
