+++ backend/src/main/java/hu/flowacademy/worksheet/dto/PartnerDTO.java
@@ -34,5 +34,4 @@ public class PartnerDTO {
     private String szamlazasiCimSzint;
     private String szamlazasiCimAjto;
     private String szamlazasiCimHrsz;
-    private Boolean enabled;
 }
