+++ backend/src/main/java/hu/flowacademy/worksheet/entity/Worksheet.java
@@ -1,8 +1,6 @@
 package hu.flowacademy.worksheet.entity;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import hu.flowacademy.worksheet.enumCustom.*;
 import hu.flowacademy.worksheet.generator.WorksheetSerialGenerator;
 import lombok.AllArgsConstructor;
@@ -38,7 +36,7 @@ public class Worksheet {
             })
     @Column(name = "worksheet_id", nullable = false)
     private String id;
-    @ManyToOne
+    @ManyToOne(cascade = CascadeType.MERGE)
     private Partner partner;
     @Enumerated(EnumType.STRING)
     @Column(name = "type_of_work", nullable = false)
