+++ backend/src/main/java/hu/flowacademy/worksheet/repository/PartnerRepository.java
@@ -3,7 +3,6 @@ package hu.flowacademy.worksheet.repository;
 import hu.flowacademy.worksheet.entity.Partner;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.stereotype.Repository;
 
 @Repository
