+++ backend/src/test/java/hu/flowacademy/worksheet/helper/TestHelper.java
@@ -38,7 +38,6 @@ public class TestHelper {
     private static final String SZINT = "II.";
     private static final String AJTO = "11";
     private static final String HRSZ = "0123-4567-8901";
-    private static final Boolean ENABLED = true;
 
     public static Header getAuthorization(String token) {
         return new Header("Authorization", "Bearer " + token);
@@ -118,7 +117,6 @@ public class TestHelper {
                 .szamlazasiCimSzint(SZINT)
                 .szamlazasiCimAjto(AJTO)
                 .szamlazasiCimHrsz(HRSZ)
-                .enabled(ENABLED)
                 .build();
     }
 }
