+++ frontend/package.json
@@ -25,7 +25,6 @@
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
-    "react-select": "^4.1.0",
     "react-signature-pad-wrapper": "^1.3.0",
     "signature_pad": "^2.3.2",
     "web-vitals": "^0.2.4",
